requirements.txt
python-telegram-bot==22.7
openai>=1.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
google-api-python-client>=2.130.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
APScheduler==3.10.4
pytz>=2024.1
